pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fitz>=0.0.1.dev2",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "streamlit>=1.48.1",
]
